refactor: move parsing of webpack __resourceQuery to where it's used
